requirements.txt
requests>=2.31.0
openai>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Notion RAG Pipeline
notion-client>=2.2.0
sqlite-vec>=0.1.1
sentence-transformers>=2.2.0
numpy>=1.24.0
